rename variables and rediscribe it test
